requirements.txt
asgiref==3.8.1
attrs==23.2.0
black==24.4.2
cffi==1.16.0
cfgv==3.4.0
click==8.1.7
cryptography==42.0.5
distlib==0.3.8
Django==5.0.4
djangorestframework==3.15.1
drf-spectacular==0.27.2
filelock==3.13.4
identify==2.5.36
inflection==0.5.1
jsonschema==4.21.1
jsonschema-specifications==2023.12.1
mypy-extensions==1.0.0
nodeenv==1.8.0
packaging==24.0
pathspec==0.12.1
platformdirs==4.2.1
pre-commit==3.7.0
pycparser==2.22
PyJWT==2.8.0
python-dotenv==1.0.1
PyYAML==6.0.1
referencing==0.35.0
rpds-py==0.18.0
sqlparse==0.5.0
uritemplate==4.1.1
virtualenv==20.26.0
